feat(offline-first): sqlite local en el dispositivo + UI autosave + sync QR claim + stack server eliminado (#60)
* feat(offline-first): sqlite local en el dispositivo + UI autosave + sync QR claim + stack server eliminado (193 unit, 12 e2e, tsc clean)

* fix(db): budgets.deleted_at en SCHEMA canónico — merge de import QR sin crash (193 unit, 12 e2e, tsc clean)

+++ tests/unit/transaction-modal.test.tsx
@@ -5,7 +5,7 @@ import { describe, it, expect, vi } from 'vitest'
 import { TransactionModal } from '@/components/modals/transaction-modal'
 import { LanguageProvider } from '@/contexts/language-context'
 import { CurrencyProvider } from '@/contexts/currency-context'
-import { Account, Int } from '@/types'
+import { Account } from '@/types'
 
 function renderWithProviders(ui: React.ReactElement) {
   return render(
@@ -17,8 +17,8 @@ function renderWithProviders(ui: React.ReactElement) {
 
 describe('TransactionModal account selector', () => {
   const accounts: Account[] = [
-    { id: 'daily', name: 'Daily Budget', type: 'daily', balance: 1000 as Int, icon: 'wallet' },
-    { id: 'savings', name: 'Savings', type: 'savings', balance: 500 as Int, icon: 'piggybank' }
+    { id: 'daily', name: 'Daily Budget', type: 'daily', balance: 1000, icon: 'wallet' },
+    { id: 'savings', name: 'Savings', type: 'savings', balance: 500, icon: 'piggybank' }
   ]
 
   function renderModal() {
@@ -63,7 +63,7 @@ describe('TransactionModal account selector', () => {
         onClose={vi.fn()}
         onAddTransaction={vi.fn()}
         onUpdateTransaction={vi.fn()}
-        accounts={[{ id: 'daily', name: 'Daily Budget', type: 'daily', balance: -300 as Int, icon: 'wallet' }]}
+        accounts={[{ id: 'daily', name: 'Daily Budget', type: 'daily', balance: -300, icon: 'wallet' }]}
         remainingToday={5000}
       />
     )
